towards unique instances of cells
Working on adding unique identifiers to cells. Right now, this breaks everything. I need to add methods that distinguish between two cells being equal (as in having the exact same data AND identifier) and that can tell if two cells have the same data (but perhaps different identifiers). The point of this is that if I want to make this a surface classifier, I need to be able to see different instances of edges that are identified in the surface (so, for example, I can find a polygonal presentation of the surface)